+++ src/dxftess-cgal.cc
@@ -106,14 +106,14 @@ void dxf_tesselate(PolySet *ps, DxfData &dxf, double rot, bool up, bool /* do_tr
 	try {
 
 	// read path data and copy all relevant infos
-	for (int i = 0; i < dxf.paths.size(); i++)
+	for (size_t i = 0; i < dxf.paths.size(); i++)
 	{
 		if (!dxf.paths[i].is_closed)
 			continue;
 
 		Vertex_handle first, prev;
 		struct point_info_t *first_pi = NULL, *prev_pi = NULL;
-		for (int j = 1; j < dxf.paths[i].indices.size(); j++)
+		for (size_t j = 1; j < dxf.paths[i].indices.size(); j++)
 		{
 			double x = dxf.points[dxf.paths[i].indices[j]][0];
 			double y = dxf.points[dxf.paths[i].indices[j]][1];
